Rename : 이미지박스 컴포넌트 수정
