+++ src/Components/ImageBox/index.js
@@ -1,0 +1,1 @@
+export { default as ImageBox } from "./ImageBox";
